# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Paper {
  id: string;
  courseCode: string;
  courseName: string;
  faculty: string;
  university: string;
  year: number;
  downloads: number;
  priceUgx: number;
  previewLabel: string;
  accent: string;
  isTrending: boolean;
  /** @nullable */
  pdfUrl: string | null;
}

export type UnlockInputMethod = typeof UnlockInputMethod[keyof typeof UnlockInputMethod];


export const UnlockInputMethod = {
  momo: 'momo',
  airtel: 'airtel',
  ad: 'ad',
} as const;

export interface UnlockInput {
  method: UnlockInputMethod;
}

export interface UnlockResult {
  unlocked: boolean;
  /** @nullable */
  downloadUrl: string | null;
  message: string;
  creditsRemaining: number;
}

export interface UploadInput {
  courseCode: string;
  courseName: string;
  university: string;
  faculty: string;
  year: number;
  fileName: string;
}

export interface UploadResult {
  id: string;
  creditsEarned: number;
  status: string;
}

export interface DashboardSummary {
  totalPapers: number;
  totalDownloads: number;
  paidUnlocks: number;
  totalEarningsUgx: number;
  pendingUploads: number;
  topUniversity: string;
}

export interface Error {
  error: string;
}

export type ListPapersParams = {
search?: string;
university?: string;
faculty?: string;
year?: number;
trending?: boolean;
};