# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalPapers: number;
  totalDownloads: number;
  paidUnlocks: number;
  totalEarningsUgx: number;
  pendingUploads: number;
  topUniversity: string;
}