# lib/api-zod/src/generated/types/paper.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Paper {
  id: string;
  courseCode: string;
  courseName: string;
  faculty: string;
  university: string;
  year: number;
  downloads: number;
  priceUgx: number;
  previewLabel: string;
  accent: string;
  isTrending: boolean;
  /** @nullable */
  pdfUrl: string | null;
}