# lib/api-zod/src/generated/types/unlockInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UnlockInputMethod } from './unlockInputMethod';

export interface UnlockInput {
  method: UnlockInputMethod;
}