# lib/api-zod/src/generated/types/uploadInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UploadInput {
  courseCode: string;
  courseName: string;
  university: string;
  faculty: string;
  year: number;
  fileName: string;
}